Kotlin android rewrite

+++ kboyemucore/src/main/java/com/github/rodrigotimoteo/kboyemucore/api/FrameBuffer.kt
@@ -15,7 +15,7 @@ import com.github.rodrigotimoteo.kboyemucore.util.WIDTH
 data class FrameBuffer(
     val width: Int = WIDTH,
     val height: Int = HEIGHT,
-    val pixels: IntArray
+    val pixels: ByteArray
 ) {
 
     override fun equals(other: Any?): Boolean {
